requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.10
attrs==25.3.0
authlib==1.6.0
blinker==1.9.0
brotli==1.1.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.0
cryptography==45.0.4
cssselect2==0.8.0
cycler==0.12.1
dill==0.4.0
docstring-parser==0.16
exceptiongroup==1.3.0
fastapi==0.115.13
flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.4
gitdb==4.0.12
gitpython==3.1.44
google-adk==1.4.1
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.97.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.21.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.73.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
isort==6.0.1
itsdangerous==2.2.0
jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
markdown==3.8.2
markdown2==2.5.3
markupsafe==3.0.2
matplotlib==3.9.4
mccabe==0.7.0
narwhals==1.43.1
networkx==3.2.1
numpy==2.0.2
opentelemetry-api==1.34.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging==25.0
pandas==2.3.0
pillow==11.2.1
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydeck==0.9.1
pydyf==0.11.0
pylint==3.3.7
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
rsa==4.9.1
shapely==2.0.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
streamlit==1.46.0
tenacity==8.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
tornado==6.5.1
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
watchdog==6.0.0
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
werkzeug==3.1.3
wordcloud==1.9.4
zipp==3.23.0
zopfli==0.2.3.post1
